prac 3

+++ PRAC/P2/C/src/CHeterodyning.c
@@ -1,9 +1,9 @@
 #include "CHeterodyning.h"
 
-extern float data [SAMPLE_COUNT];
-extern float carrier[SAMPLE_COUNT];
+extern __fp16 data [SAMPLE_COUNT];
+extern __fp16 carrier[SAMPLE_COUNT];
 
-float result [SAMPLE_COUNT];
+__fp16 result [SAMPLE_COUNT];
 
 int main(int argc, char**argv){
     printf("Running Unthreaded Test\n");
@@ -17,7 +17,7 @@ int main(int argc, char**argv){
     for (int i = 0;i<SAMPLE_COUNT;i++ ){
         result[i] = data[i] * carrier[i];
     }
-    double t = toc();
+    __fp16 t = toc();
     printf("Time: %lf ms\n",t/1e-3);
     printf("End Unthreaded Test\n");
     return 0;
